+++ src/manager/collectionManager/CollectionManager.vue
@@ -3,7 +3,7 @@
     <v-expansion-panels
       accordion
       multiple
-      v-if="componentIds.length > 0"
+      v-if="!componentView && componentIds.length > 0"
       v-model="panels"
       class="rounded-0"
     >
@@ -11,8 +11,16 @@
         v-for="componentId in componentIds"
         :component-id="componentId"
         :key="componentId"
-      />
+      >
+        <collection-component @openList="openList" />
+      </collection-component-provider>
     </v-expansion-panels>
+    <collection-component-provider
+      v-if="componentView"
+      :component-id="componentView"
+    >
+      <collection-component-list @closeList="closeList" />
+    </collection-component-provider>
   </v-container>
 </template>
 
@@ -20,6 +28,8 @@
   import { computed, inject, ref } from 'vue';
   import { VExpansionPanels, VContainer } from 'vuetify/lib';
   import CollectionComponentProvider from './CollectionComponentProvider.vue';
+  import CollectionComponentList from './CollectionComponentList.vue';
+  import CollectionComponent from './CollectionComponent.vue';
 
   /**
    * @description Renders the all managed CollectionComponents of a CollectionManager.
@@ -28,16 +38,24 @@
   export default {
     name: 'CollectionManager',
     components: {
+      CollectionComponentList,
       VExpansionPanels,
       VContainer,
       CollectionComponentProvider,
+      CollectionComponent,
     },
     setup() {
+      /**
+       * @type {import("./collectionManager.js").CollectionManager}
+       */
       const collectionManager = inject('collectionManager');
       const componentIds = ref(collectionManager.componentIds);
-
       /**
-       * @type {WritableComputedRef<number[]>}
+       * @type {import("vue").Ref<string|null>}
+       */
+      const componentView = ref(null);
+      /**
+       * @type {import("vue").WritableComputedRef<number[]>}
        */
       const panels = computed({
         get() {
@@ -56,6 +74,13 @@
       return {
         componentIds,
         panels,
+        componentView,
+        openList(id) {
+          componentView.value = id;
+        },
+        closeList() {
+          componentView.value = null;
+        },
       };
     },
   };